pyproject.toml
[project]
name = "kurikone-cb-bot"
version = "0.1.0"
description = "Princess Connect Discord CB Bot"
requires-python = ">=3.12"
dependencies = [
    "attrs==25.3.0",
    "discord-py==2.5.2",
    "dotenv>=0.9.9",
    "mariadb>=1.1.12",
]
